% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BessFix.R
\name{BessFix}
\alias{BessFix}
\title{BessFix}
\usage{
BessFix(s, x, y, kmax, TAWs, n, p, corrsY)
}
\arguments{
\item{s}{an integer, the maximum support size of coefficient we allow, given and fixed}

\item{x}{the predictors of observations}

\item{y}{the responses of observations}

\item{kmax}{an integer representing the maximum splicing size}

\item{TAWs}{a numeric value representing the threshold chosen by user into the splicing method}

\item{n}{an integer representing number of observations}

\item{p}{an integer representing dimensions of regression coefficient vector}

\item{corrsY}{correlations imported into BessFix in order to help determine the initial value for setA_0}
}
\description{
this function is for best subset selection with support size s given and fixed.
}
