% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2NormSq.R
\name{l2NormSq}
\alias{l2NormSq}
\title{l2 Norm Sq}
\usage{
l2NormSq(b)
}
\arguments{
\item{b}{a vector}
}
\description{
we use this function to calculate the square of l2 norm of a vector.
}
\examples{
\dontrun{
x = c(1:3)
print(l2NormSq(x))
}
}
