% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matmultC}
\alias{matmultC}
\title{matrix multiplication}
\usage{
matmultC(matA, matB)
}
\arguments{
\item{matA}{a matrix}

\item{matB}{another matrix with which matA can do a matrix multiplication}
}
\value{
The result C of C=AB
}
\description{
Compute the result of matrix multiplication in cpp
}
\examples{
\dontrun{
mat1 = matrix(rnorm(15),nrow = 3,ncol = 5)
mat2 = matrix(rnorm(30),nrow = 5,ncol = 6)
c = matmultC(mat1,mat2)
print(c)
}
}
