% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myABESS.R
\name{myABESS}
\alias{myABESS}
\title{abess}
\usage{
myABESS(x, y, n, p, smax = min(p, floor(n/log(p)/log(log(n)))))
}
\arguments{
\item{x}{the input x, predictors of observations}

\item{y}{the input y, responses of observations}

\item{n}{an integer representing number of observations}

\item{p}{an integer representing dimensions of regression coefficient vector}

\item{smax}{the maximum support size that can be allowed, default set to floor( n/log(p)/log(log(n)) )}
}
\description{
the function for adaptive best-subset selection.
}
\examples{
\dontrun{
library(mvtnorm)
set.seed(123)
n=60
eps_sig=1
p = 20
beta= c(0,0,10, rep(0,p-3))
sig=matrix(0, nrow=p,ncol=p)
sig= 0.8^( abs( row(sig)-col(sig) ) )

x=rmvnorm(n,mean=rep(0,p),sigma=sig)
x=scale(x,center = T,scale = F)
y = rnorm(n, mean=x \%*\% beta, eps_sig)

x=as.matrix(x)
y=as.matrix(y)
syn_sim_data = list( "x"=x,"y"=y,"beta"=beta )
dat = cbind.data.frame("y" = syn_sim_data[["y"]],syn_sim_data[["x"]])
myABESS(x,y,n,p,smax = p-1)
}
}
