% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splicing.R
\name{Splicing}
\alias{Splicing}
\title{Splicing}
\usage{
Splicing(beta, d, setA, setI, kmax, TAWs, n, p, x, y)
}
\arguments{
\item{beta}{the regression coefficient vector that is imported into the splicing method}

\item{d}{the vector value of t(x)(y-x beta)/n that is imported into the splicing method}

\item{setA}{a vector representing the set of positions of non zero values of regression coefficient that is imported into the splicing method}

\item{setI}{a vector representing the set of positions of zero values of regression coefficient that is imported into the splicing method}

\item{kmax}{an integer representing the maximum splicing size}

\item{TAWs}{a numeric value representing the threshold chosen by user into the splicing method}

\item{n}{an integer representing number of observations}

\item{p}{an integer representing dimensions of regression coefficient vector}

\item{x}{the predictors of observations}

\item{y}{the responses of observations}
}
\description{
splicing method for adaptive best subset selection.
}
